agent/instagram_bot/requirements.txt
python-telegram-bot
sqlalchemy
python-dotenv
instagrapi
Pillow>=8.1.1
openai
httpx
